# improve performance of DGMulti methods using entropy-projections (for 2D Euler)