--- a/src/solvers/dgmulti.jl
+++ b/src/solvers/dgmulti.jl
@@ -2,3 +2,6 @@
 include("dgmulti/types.jl")
 include("dgmulti/dg.jl")
 include("dgmulti/flux_differencing.jl")
+
+# specialization of DGMulti to specific equations
+include("dgmulti/flux_differencing_compressible_euler.jl")