@sloede 2021-08-24
> CC @sloede for approving the general strategy of performance optimizations specific to combinations of the solver/equations.

Personally, I think the code is rather difficult to understand and hard to maintain by anyone else but the initiated, but I do see the benefits of the performance improvements. As long as this is not getting out of hand and spreading to the rest of Trixi, I'd say go for it :shrug: :sunglasses: