@ranocha 2021-08-24
My [second PR to StructArrays.jl](https://github.com/JuliaArrays/StructArrays.jl/pull/196) speeds up weak forms stuff such as `trixi_include("examples/dgmulti_2d/elixir_euler_triangulate_pkg_mesh.jl")` by ca. one third on top of the other performance improvements mentioned above.