@ranocha 2021-08-24
Thanks for your nice review, @jlchan!

> This is very cool, and gives me some ideas for my next PR. 

Just don't push the specializations too hard or @sloede will kill us (and Trixi.jl will become incredibly unreadable).

> It's also a nice example of "extra" specialization that can be done on top of Trixi.

Yeah - this is a kind of specialization that I prototyped in https://github.com/trixi-framework/Trixi.jl/pull/429, where we were able to beat some MFEM variants for linear advection :tada: It's a really nice feature of Julia