@jlchan 2021-08-24
> Just don't push the specializations too hard or @sloede will kill us (and Trixi.jl will become incredibly unreadable).

Yeah - that's one thing I'm worried about. This addresses the fact that `entropy_projection!` is slow, but the solver that it makes the most impact for (e.g., `DGMultiFluxDiff{<:Polynomial, Quad}`) has some inefficiencies that can be eliminated. A change of basis from nodal bases at Lobatto nodes to Gauss nodes exposes structure in `entropy_projection!` that can be used to eliminate all the volume evaluations of `entropy2cons` and `cons2entropy`. 

This PR is helpful to speed things up until we can introduce a `DGMultiFluxDiff{<:GSBP, Quad}` solver type. When that gets implemented, I'd suggest that we revisit `flux_differencing_compressible_euler.jl` and maybe move it to an elixir or decide what solver type it should be specialized for.