@sloede 2021-08-24
> > Just don't push the specializations too hard or @sloede will kill us (and Trixi.jl will become incredibly unreadable).
> 
> Yeah - that's one thing I'm worried about.

About what exactly - the readability issue or about getting killed? Please tell me it's the latter...