@ranocha 2021-08-24
> This PR is helpful to speed things up until we can introduce a `DGMultiFluxDiff{<:GSBP, Quad}` solver type. When that gets implemented, I'd suggest that we revisit `flux_differencing_compressible_euler.jl` and maybe move it to an elixir or decide what solver type it should be specialized for.

Sounds good to me. The problems I addressed in my PRs to StructArrays.jl will still be helpful either way